Bills in CoreData - Eliminating some transformables
Had to restructure the bills model, reducing the number of transformable properties (in favor of relationships) in order to allow the managed object cache (and anyone else) to perform searches on subjects, types, alternate_titles, etc.  In the process, we refactored the bill votes and voters too.  There are still transformable properties around, but not super important ones.